requirements.txt
python-binance==1.0.29
python-dotenv==1.1.1
pydantic==2.11.7
rich==14.1.0
